Migrate from IsrPointer to std::function
Split usage of IsrPointer into two different categories of interrupt
callables.

For actual interrupts that get put into the IVT, such as an interrupt
service routine, they will be named `InterruptHandler`.
For calling a callable objects from within an interrupt service routine,
these will be called `InterruptCallback`.

`InterruptHanlder` is defined as a function pointer of void(*)(void),
where as `InterruptCallback` a std::function<void(void)>, which allows
the usage of function pointers as well as capturing lambdas, bindings
to class methods, and usage of functors.

Caveats to this is that the <functional> header increases compile time
and occupies 32 bytes of space per object, which is 4x the size of a
pointer.

Various fixes found when migrating to ARM GCC 9

Resolves #952

+++ library/L2_HAL/sensors/movement/accelerometer/mma8452q.hpp
@@ -58,7 +58,7 @@ class Mma8452q : public Accelerometer
                         sizeof(tilt_val));
     tilt_reading = (tilt_val[0] << kMsbShift) | tilt_val[1];
     axis_tilt    = static_cast<int16_t>(tilt_reading);
-    return axis_tilt / kDataOffset;
+    return static_cast<int16_t>(axis_tilt / kDataOffset);
   }
   int16_t X() const override
   {
